+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Download facilities list as Excel [#1293](https://github.com/open-apparel-registry/open-apparel-registry/pull/1293)
+- Add scripts/console [#1295](https://github.com/open-apparel-registry/open-apparel-registry/pull/1295)
 
 ### Changed