+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Remove Sample Data and Update Registration Text [#1291](https://github.com/open-apparel-registry/open-apparel-registry/pull/1291)
+- Update Dependencies [#1291](https://github.com/open-apparel-registry/open-apparel-registry/pull/1303)
 
 ### Deprecated